✏️ #38 Rename: 파일명, 폴더 구조 변경
