+++ src/main/java/kr/co/teacherforboss/config/jwt/exception/JwtAuthenticationEntryPoint.java
@@ -7,33 +7,33 @@ import jakarta.servlet.http.HttpServletResponse;
 import kr.co.teacherforboss.apiPayload.ApiResponse;
 import kr.co.teacherforboss.apiPayload.code.status.ErrorStatus;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
 @Slf4j
 @Component
-public class MyAccessDeniedHandler implements AccessDeniedHandler {
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     @Override
-    public void handle(HttpServletRequest request,
-                       HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException, ServletException {
+    public void commence(HttpServletRequest request,
+                         HttpServletResponse response,
+                         AuthenticationException authException) throws IOException, ServletException {
 
         if (request.getRequestURI().startsWith("/api/v1/auth/")) {
-            log.info("[접근 권한 X] URL : [{}]", request.getRequestURL());
+            log.info("[로그인 X] MyAuthenticationEntryPoint");
 
             ObjectMapper objectMapper = new ObjectMapper();
-            String code = ErrorStatus.ACCESS_DENIED.getCode();
-            String message = ErrorStatus.ACCESS_DENIED.getMessage();
+            String code = ErrorStatus.INVALID_JWT_TOKEN.getCode();
+            String message = ErrorStatus.INVALID_JWT_TOKEN.getMessage();
 
             ApiResponse<ErrorStatus> apiResponse = ApiResponse.onFailure(code, message, null);
             String jsonResponse = objectMapper.writeValueAsString(apiResponse);
 
             response.setContentType("application/json;charset=UTF-8");
-            response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
             response.getWriter().write(jsonResponse);
         }
     }
