+++ src/main/java/kr/co/teacherforboss/config/jwt/filter/JwtAuthenticationFilter.java
@@ -1,9 +1,10 @@
-package kr.co.teacherforboss.config.jwt;
+package kr.co.teacherforboss.config.jwt.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import kr.co.teacherforboss.config.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
