+++ src/main/java/kr/co/teacherforboss/config/jwt/filter/JwtExceptionFilter.java
@@ -1,4 +1,4 @@
-package kr.co.teacherforboss.config.jwt;
+package kr.co.teacherforboss.config.jwt.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
